requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==4.0.6
django-cors-headers==3.8.0
django-dotenv==1.4.2
djangorestframework==3.12.4
drf-yasg==1.20.0
flake8==4.0.1
gunicorn==20.1.0
idna==3.2
mccabe==0.6.1
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
